+++ src/element.ts
@@ -61,8 +61,12 @@ export abstract class Element {
   protected boundingBox?: BoundingBox;
   protected registry?: Registry;
 
-  /** Some elements include text. You can customize the font family, size, weight, and style. */
-  protected font?: Required<FontInfo>;
+  /**
+   * Some elements include text.
+   * The `textFont` property contains information required to style the text (i.e., font family, size, weight, and style).
+   * It starts as `undefined`, and must be set using `setFont()` or `resetFont()`.
+   */
+  protected textFont?: Required<FontInfo> = undefined;
 
   constructor() {
     this.attrs = {
@@ -84,40 +88,65 @@ export abstract class Element {
   }
 
   /**
-   * Return the default text font. To render text, an element should
-   * call `this.setFont(this.getDefaultFont())` in its constructor.
-   */
-  getDefaultFont(): Required<FontInfo> {
-    return (<typeof Element>this.constructor).TEXT_FONT;
-  }
-
-  /**
    * Set the element's font family, size, weight, style (e.g., `Arial`, `10pt`, `bold`, `italic`).
-   *
-   * @param f a string that specifies the font family, or a `FontInfo` options object.
-   * If the first argument is a `FontInfo`, the other arguments below are ignored.
+   * @param f is 1) a `FontInfo` object or
+   *             2) a string formatted as CSS font shorthand (e.g., 'bold 10pt Arial') or
+   *             3) a string representing the font family (one of `size`, `weight`, or `style` must also be provided).
    * @param size a string specifying the font size and unit (e.g., '16pt'), or a number (the unit is assumed to be 'pt').
-   * @param weight is inserted into the font-weight attribute (e.g., font-weight="bold")
-   * @param style is inserted into the font-style attribute (e.g., font-style="italic")
+   * @param weight is a string (e.g., 'bold', 'normal') or a number (100, 200, ... 900).
+   * @param style is a string (e.g., 'italic', 'normal').
+   * If no arguments are provided, then the font is set to the default font.
+   * Each Element subclass may specify its own default by overriding the static `TEXT_FONT` property.
    */
-  setFont(
-    f: string | FontInfo = Font.SANS_SERIF,
-    size: string | number = Font.SIZE,
-    weight: string | number = FontWeight.NORMAL,
-    style: string = FontStyle.NORMAL
-  ): this {
-    if (typeof f === 'string') {
-      this.font = { family: f, size, weight, style };
+  setFont(f?: string | FontInfo, size?: string | number, weight?: string | number, style?: string): this {
+    // Allow subclasses to override `TEXT_FONT`.
+    const defaultTextFont = (<typeof Element>this.constructor).TEXT_FONT;
+    const sizeWeightStyleUndefined = size === undefined && weight === undefined && style === undefined;
+    if (typeof f === 'object' || (f === undefined && sizeWeightStyleUndefined)) {
+      // `f` is case 1) a FontInfo object, or all arguments are undefined.
+      // Do not check `arguments.length === 0`, to allow the extreme edge case:
+      // setFont(undefined, undefined, undefined, undefined).
+      this.textFont = { ...defaultTextFont, ...f };
+    } else if (typeof f === 'string' && sizeWeightStyleUndefined) {
+      // `f` is case 2) CSS font shorthand.
+      this.textFont = Font.fromCSSString(f);
     } else {
-      // Follow CSS conventions. Unspecified params are reset to the default.
-      this.font = { ...this.getDefaultFont(), ...f };
+      // `f` is case 3) a font family string (e.g., 'Times New Roman')
+      // or it is undefined while one or more of the other arguments is provided.
+      // Following CSS conventions, any unspecified params are reset to the default.
+      this.textFont = Font.validate(
+        f ?? defaultTextFont.family,
+        size ?? defaultTextFont.size,
+        weight ?? defaultTextFont.weight,
+        style ?? defaultTextFont.style
+      );
     }
     return this;
   }
 
+  /**
+   * Reset the text font to the style indicated by the static `TEXT_FONT` property.
+   * Subclasses can call this to initialize `textFont` for the first time.
+   */
+  resetFont(): void {
+    this.setFont();
+  }
+
+  getFont(): string {
+    return Font.toCSSString(this.textFont);
+  }
+
   /** Return a copy of the FontInfo object, or undefined if `setFont()` has never been called. */
-  getFont(): Required<FontInfo> | undefined {
-    return this.font ? { ...this.font } : undefined;
+  getFontInfo(): Required<FontInfo> | undefined {
+    return this.textFont ? { ...this.textFont } : undefined;
+  }
+
+  set font(f: string) {
+    this.setFont(f);
+  }
+
+  get font(): string {
+    return Font.toCSSString(this.textFont);
   }
 
   /** Set the draw style of a stemmable note. */